backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
bcrypt==4.1.1
pyjwt==2.8.1
joblib==1.3.2
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
python-multipart==0.0.6
